saga_gui, map/messages control: improved dark mode support and font scaling

+++ saga-gis/src/saga_core/saga_gui/wksp_map.cpp
@@ -931,7 +931,7 @@ CWKSP_Map_Layer * CWKSP_Map::Add_Layer(CWKSP_Layer *pLayer)
 		{
 			m_Name = pLayer->Get_Object()->Get_Name();
 
-			if( m_Name.Length() > g_pMaps->Get_Parameter("NAMING_LENGTH")->asInt() )
+			if( m_Name.Length() > (size_t)g_pMaps->Get_Parameter("NAMING_LENGTH")->asInt() )
 			{
 				m_Name = m_Name.Left(g_pMaps->Get_Parameter("NAMING_LENGTH")->asInt() - 3) + "...";
 			}
